refactor: main함수 파일명 변경
